change: ffmpeg service logic to incorporate separate logic for conversion and resize
