+++ src/main/java/cloud/ptl/povserver/ffmpeg/FfmpegService.java
@@ -2,42 +2,28 @@ package cloud.ptl.povserver.ffmpeg;
 
 // ffmpeg -i 'I made Brett do this....webm' -s 1280x720 -vcodec libx265 'I made Brett do this...1280x720.mp4'
 
-import cloud.ptl.povserver.data.model.ResolutionDAO;
 import cloud.ptl.povserver.data.model.ResourceDAO;
-import cloud.ptl.povserver.service.resource.ResolutionService;
-import cloud.ptl.povserver.service.resource.ResourceService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.data.util.Pair;
+import cloud.ptl.povserver.ffmpeg.convert.ConvertRequest;
+import cloud.ptl.povserver.ffmpeg.convert.FormatConverter;
+import cloud.ptl.povserver.ffmpeg.resize.ResizeRequest;
+import cloud.ptl.povserver.ffmpeg.resize.ResizeService;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
-import javax.imageio.ImageIO;
-import javax.imageio.ImageReader;
-import javax.imageio.stream.ImageInputStream;
-import java.io.BufferedReader;
-import java.io.File;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.util.List;
 
 /**
  * Service is used to call external program to convert vide into other formats.
- * ffmpeg -i earth.gif -c:v libvpx-vp9 -crf 30 -b:v 0 -b:a 128k -c:a libopus -s 476x280 out.webm
  */
 @Service
+@Slf4j
 public class FfmpegService {
-    private final String FFMPEG_RESIZE_COMMAND =
-            "/usr/bin/ffmpeg -i %s -c:v libvpx-vp9 -crf 30 -b:v 0 -b:a 128k -c:a libopus -s %dx%d %s%dx%d.webm";
-    private final String FFMPEG_CONVERT_COMMAND =
-            "/usr/bin/ffmpeg -i %s -c:v libvpx-vp9 -crf 30 -b:v 0 -b:a 128k -c:a libopus -s %dx%d %s.webm";
+    private final ResizeService resizeService;
+    private final FormatConverter formatConverter;
 
-    private final Logger logger = LoggerFactory.getLogger(FfmpegService.class);
-    private final ResourceService resourceService;
-    private final ResolutionService resolutionService;
-
-    public FfmpegService(ResourceService resourceService, ResolutionService resolutionService) {
-        this.resourceService = resourceService;
-        this.resolutionService = resolutionService;
+    public FfmpegService(ResizeService resizeService, FormatConverter formatConverter) {
+        this.resizeService = resizeService;
+        this.formatConverter = formatConverter;
     }
 
     /**
@@ -48,145 +34,11 @@ public class FfmpegService {
      * @throws IOException if file cannot be found
      */
     public ResourceDAO resize(ResizeRequest resizeRequest) throws IOException, InterruptedException {
-        List<ResolutionDAO> availableResolutions = resizeRequest.getResourceDAO().getResolutions();
-        // checking if conversion was made before
-        boolean alreadyContains =
-                availableResolutions.stream()
-                        .anyMatch(el ->
-                                el.getHeight() == resizeRequest.getHeight() && el.getWidth() == resizeRequest.getWidth()
-                        );
-        if (!alreadyContains) {
-            this.callFFMPEGtoResize(resizeRequest);
-        }
-        // update path to newly created resource
-        String newPath = resizeRequest.getResourceDAO().getMovie().getAbsolutePath();
-        newPath += resizeRequest.getWidth() + "x" + resizeRequest.getHeight() + ".webm";
-        ResourceDAO resizedResource = resizeRequest.getResourceDAO();
-        resizedResource.setMovie(new File(newPath));
-        return resizedResource;
+        return this.resizeService.resize(resizeRequest);
     }
 
-    /**
-     * Calls FFMPEG command line utility to generate resized resource
-     *
-     * @param resizeRequest request containing all needed informations to conduct resizing
-     * @throws IOException thrown if resize cannot be done, because file was not found
-     */
-    private void callFFMPEGtoResize(ResizeRequest resizeRequest) throws IOException, InterruptedException {
-        String inflateCommand =
-                String.format(
-                        this.FFMPEG_RESIZE_COMMAND,
-                        resizeRequest.getResourceDAO().getMovie().getAbsolutePath(),
-                        resizeRequest.getWidth(),
-                        resizeRequest.getHeight(),
-                        resizeRequest.getResourceDAO().getMovie().getAbsolutePath(),
-                        resizeRequest.getWidth(),
-                        resizeRequest.getHeight()
-                );
-        ProcessBuilder processBuilder = new ProcessBuilder(inflateCommand.split(" "));
-        processBuilder.redirectErrorStream(true);
-        Process process = processBuilder.start();
-        // connect to process streams
-        BufferedReader outputBufferedReader = new BufferedReader(new InputStreamReader(process.getInputStream()));
-        BufferedReader errorBufferedReader = new BufferedReader(new InputStreamReader(process.getErrorStream()));
-
-        String line;
-
-        while ((line = outputBufferedReader.readLine()) != null) {
-            this.logger.info(line);
-        }
-        while ((line = errorBufferedReader.readLine()) != null) {
-            this.logger.info(line);
-        }
-
-        process.waitFor();
-        if (process.exitValue() == 0) {
-            // processed successfully
-            // rewrite data to resource
-            ResolutionDAO resolutionDAO = new ResolutionDAO();
-            resolutionDAO.setWidth(resizeRequest.getWidth());
-            resolutionDAO.setHeight(resizeRequest.getHeight());
-            resolutionDAO = this.resolutionService.save(resolutionDAO);
-            resizeRequest.getResourceDAO().getResolutions().add(resolutionDAO);
-            resizeRequest.setResourceDAO(
-                    this.resourceService.save(
-                            resizeRequest.getResourceDAO()
-                    )
-            );
-        } else {
-            // some error occurred
-            throw new RuntimeException("Cannot convert file to given resolution, ffmpeg exit code in none-zero");
-        }
+    public ResourceDAO convert(ConvertRequest convertRequest) throws IOException, InterruptedException {
+        return this.formatConverter.convert(convertRequest);
     }
 
-    public ResourceDAO convertGifToWebM(ConvertRequest convertRequest) throws IOException, InterruptedException {
-        Pair<Integer, Integer> dimensions = this.getGifDimensions(convertRequest.getFileToConvert());
-        this.callFFMPEGtoConvertToGif(convertRequest, dimensions);
-        ResolutionDAO resolutionDAO = new ResolutionDAO();
-        resolutionDAO.setHeight(dimensions.getFirst());
-        resolutionDAO.setWidth(dimensions.getSecond());
-        resolutionDAO = this.resolutionService.save(resolutionDAO);
-
-        ResourceDAO newResourceDAO = new ResourceDAO();
-        newResourceDAO.setResolutions(List.of(resolutionDAO));
-        newResourceDAO.setMovie(
-                new File(
-                        convertRequest.getDestinationFolder().getAbsolutePath() + File.separator + convertRequest.getFileToConvert().getName() + ".webm"
-                )
-        );
-        newResourceDAO.setTitle(convertRequest.getFileToConvert().getName());
-        return this.resourceService.save(newResourceDAO);
-    }
-
-    private Pair<Integer, Integer> getGifDimensions(File file) throws IOException {
-        ImageReader is = ImageIO.getImageReadersBySuffix("GIF").next();
-        ImageInputStream iis;
-        try {
-            iis = ImageIO.createImageInputStream(file);
-            is.setInput(iis);
-            return Pair.of(
-                    is.getWidth(0),
-                    is.getHeight(0)
-            );
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return null;
-    }
-
-    private void callFFMPEGtoConvertToGif(ConvertRequest convertRequest, Pair<Integer, Integer> dimensions) throws IOException, InterruptedException {
-        String inflateCommand =
-                String.format(
-                        this.FFMPEG_CONVERT_COMMAND,
-                        convertRequest.getFileToConvert().getAbsolutePath(),
-                        dimensions.getFirst(),
-                        dimensions.getSecond(),
-                        convertRequest.getDestinationFolder().getAbsolutePath() + File.separator + convertRequest.getFileToConvert().getName()
-                );
-        ProcessBuilder processBuilder = new ProcessBuilder(inflateCommand.split(" "));
-        processBuilder.redirectErrorStream(true);
-        Process process = processBuilder.start();
-        // connect to process streams
-        BufferedReader outputBufferedReader = new BufferedReader(new InputStreamReader(process.getInputStream()));
-        BufferedReader errorBufferedReader = new BufferedReader(new InputStreamReader(process.getErrorStream()));
-
-        String line;
-
-        while ((line = outputBufferedReader.readLine()) != null) {
-            this.logger.info(line);
-        }
-        while ((line = errorBufferedReader.readLine()) != null) {
-            this.logger.info(line);
-        }
-
-        process.waitFor();
-        if (process.exitValue() == 0) {
-            // processed successfully
-            // rewrite data to resource
-            return;
-        } else {
-            // some error occurred
-            throw new RuntimeException("Cannot convert file to given resolution, ffmpeg exit code in none-zero");
-        }
-    }
 }
