+++ src/main/java/cloud/ptl/povserver/service/stream/StreamService.java
@@ -3,7 +3,7 @@ package cloud.ptl.povserver.service.stream;
 import cloud.ptl.povserver.data.model.ResourceDAO;
 import cloud.ptl.povserver.exception.NotFoundException;
 import cloud.ptl.povserver.ffmpeg.FfmpegService;
-import cloud.ptl.povserver.ffmpeg.ResizeRequest;
+import cloud.ptl.povserver.ffmpeg.resize.ResizeRequest;
 import cloud.ptl.povserver.service.metric.MetricsService;
 import cloud.ptl.povserver.service.resource.ResourceService;
 import org.apache.commons.lang3.StringUtils;
